test_samples/weather-agent-framework/requirements.txt
# M365 Agents SDK for Teams/M365 integration
#microsoft-agents-hosting-aiohttp
#microsoft-agents-authentication-msal  # uncomment to enable MSAL-based auth

# Azure services
azure-identity

# OpenAI
openai

# Web framework
aiohttp

# Configuration
python-dotenv
pydantic

# Weather API client
pyowm

# Utilities
typing-extensions

# OpenTelemetry — core (required for telemetry/)
opentelemetry-api
opentelemetry-sdk

# OpenTelemetry — exporters
# OTLP (traces, metrics, logs) — for Aspire dashboard, Jaeger, etc.
opentelemetry-exporter-otlp
# Azure Monitor / Application Insights
azure-monitor-opentelemetry-exporter

# OpenTelemetry — library auto-instrumentation
# aiohttp server (inbound) and client (outbound)
opentelemetry-instrumentation-aiohttp-server
opentelemetry-instrumentation-aiohttp-client
# requests library (outbound HTTP)
opentelemetry-instrumentation-requests
# Correlate Python log records with trace context
opentelemetry-instrumentation-logging

# A365 observability — AgentFramework automatic instrumentation
# Provides AgentFrameworkInstrumentor used in telemetry/agent_otel_extensions.py
microsoft-agents-a365-runtime==0.2.1.dev33
microsoft-agents-a365-observability-core==0.2.1.dev33
microsoft-agents-a365-observability-extensions-agent-framework>=0.2.1.dev32
